+++ packages/core/breadcrumbs/src/components/EllipsisItem.tsx
@@ -6,6 +6,8 @@ import Separator from '../styled/Separator';
 interface IProps {
   hasSeparator?: boolean;
   onClick?: (event: React.MouseEvent) => any;
+  /** A `testId` prop is provided for specified elements, which is a unique string that appears as a data attribute `data-testid` in the rendered code, serving as a hook for automated tests */
+  testId?: string;
 }
 
 export default class EllipsisItem extends React.Component<IProps, {}> {
@@ -15,11 +17,13 @@ export default class EllipsisItem extends React.Component<IProps, {}> {
   };
 
   render() {
+    const { testId } = this.props;
     return (
       <ItemWrapper>
         <Button
           appearance="subtle-link"
           spacing="none"
+          testId={testId}
           onClick={this.props.onClick}
         >
           &hellip;