+++ packages/help/help/examples/1-Help-Default-Content.tsx
@@ -10,7 +10,12 @@ import {
   FooterContent,
   ExampleDefaultContent,
 } from './utils/styled';
-import Help, { ArticleFeedback } from '../src';
+import Help from '../src';
+
+interface ArticleFeedback {
+  RateReasonText: string;
+  negativeRateReason?: string;
+}
 
 const handleEvent = (analyticsEvent: { payload: any; context: any }) => {
   const { payload, context } = analyticsEvent;