+++ packages/help/help/src/components/Article/ArticleListItem/index.tsx
@@ -4,6 +4,7 @@ import {
   CreateUIAnalyticsEvent,
   UIAnalyticsEvent,
 } from '@atlaskit/analytics-next';
+import DocumentFilledIcon from '@atlaskit/icon/glyph/document-filled';
 import ShortcutIcon from '@atlaskit/icon/glyph/shortcut';
 
 import {
@@ -22,24 +23,36 @@ import {
   ArticlesListItemLinkIcon,
 } from './styled';
 
-interface Props {
+type Props = {
+  /* Analytics event */
   createAnalyticsEvent: CreateUIAnalyticsEvent;
-  onClick: (id: string, analyticsEvent: UIAnalyticsEvent) => void;
-  title: string;
-  description: string;
-  icon: React.ReactNode;
+  /* Function executed when the user clicks the related article */
+  onClick?: (id: string, analyticsEvent: UIAnalyticsEvent) => void;
+  /* Related article title. This prop is optional (default value is '') */
+  title?: string;
+  /* Related article description. This prop is optional (default value is '') */
+  description?: string;
+  /* Related article icon. This prop is optional (by default a DocumentFilledIcon is used) */
+  icon?: React.ReactNode;
+  /* Related article href. This prop is optional (default value is ''). If is defined, when 
+  the user clicks in the related article a new tab will be open using the url defined in this prop */
   href?: string;
+  /* Related article ID */
   id: string;
-}
+};
 
-const ArticlesListItem = (props: Props & Analytics) => {
+const ArticlesListItem: React.SFC<Props & Analytics> = (
+  props: Props & Analytics,
+) => {
   const {
     id,
-    title,
-    description,
-    icon,
-    onClick,
-    href,
+    title = '',
+    description = '',
+    icon = (
+      <DocumentFilledIcon primaryColor={colors.P300} size="medium" label="" />
+    ),
+    href = '',
+    onClick = (id: string, analyticsEvent: UIAnalyticsEvent) => {},
     createAnalyticsEvent,
   } = props;