+++ packages/navigation/atlassian-switcher/src/components/generic-switcher.tsx
@@ -27,13 +27,11 @@ type GenericSwitcherProps = WithTheme & {
 
 export default (props: GenericSwitcherProps) => (
   <AvailableProductsProvider
-    isUserCentric={props.features.enableUserCentricProducts}
     availableProductsDataProvider={props.availableProductsDataProvider}
   >
     {availableProducts => (
       <CommonDataProvider
         cloudId={props.cloudId}
-        isUserCentric={Boolean(props.features.enableUserCentricProducts)}
         disableRecentContainers={props.features.disableRecentContainers}
       >
         {providerResults => {
@@ -44,7 +42,6 @@ export default (props: GenericSwitcherProps) => (
             availableProducts,
             props.product,
           );
-
           return <Switcher {...props} {...switcherLinks} />;
         }}
       </CommonDataProvider>