+++ packages/navigation/atlassian-switcher/src/utils/links.tsx
@@ -9,7 +9,6 @@ import MarketplaceGlyph from '@atlaskit/icon/glyph/marketplace';
 import {
   BitbucketIcon,
   ConfluenceIcon,
-  JiraIcon,
   JiraSoftwareIcon,
   JiraServiceDeskIcon,
   JiraCoreIcon,
@@ -19,8 +18,6 @@ import {
 } from '@atlaskit/logo';
 import FormattedMessage from '../primitives/formatted-message';
 import {
-  LicenseInformationResponse,
-  ProductLicenseInformation,
   RecentContainerType,
   AvailableProductsResponse,
   AvailableProduct,
@@ -28,20 +25,13 @@ import {
   ProductKey,
   RecommendationsEngineResponse,
   Product,
+  ProvisionedProducts,
+  CurrentSite,
 } from '../types';
 import messages from './messages';
-import JiraOpsLogo from './assets/jira-ops-logo';
-import PeopleLogo from './assets/people';
 import { CustomLink, RecentContainer, SwitcherChildItem } from '../types';
 import WorldIcon from '@atlaskit/icon/glyph/world';
 import { createIcon, createImageIcon, IconType } from './icon-themes';
-
-enum ProductActivationStatus {
-  ACTIVE = 'ACTIVE',
-  DEACTIVATED = 'DEACTIVATED',
-}
-
-const SINGLE_JIRA_PRODUCT: 'jira' = 'jira';
 
 interface MessagesDict {
   [index: string]: FormattedMessageNamespace.MessageDescriptor;
@@ -67,50 +57,6 @@ export const OBJECT_TYPE_TO_LABEL_MAP: MessagesDict = {
   'confluence-space': messages.confluenceSpace,
 };
 
-export const PRODUCT_DATA_MAP: {
-  [productKey in ProductKey | typeof SINGLE_JIRA_PRODUCT]: {
-    label: string;
-    Icon: React.ComponentType<any>;
-    href: string;
-  }
-} = {
-  [ProductKey.CONFLUENCE]: {
-    label: 'Confluence',
-    Icon: createIcon(ConfluenceIcon, { size: 'small' }),
-    href: '/wiki',
-  },
-  [ProductKey.JIRA_CORE]: {
-    label: 'Jira Core',
-    Icon: createIcon(JiraCoreIcon, { size: 'small' }),
-    href: '/secure/BrowseProjects.jspa?selectedProjectType=business',
-  },
-  [ProductKey.JIRA_SOFTWARE]: {
-    label: 'Jira Software',
-    Icon: createIcon(JiraSoftwareIcon, { size: 'small' }),
-    href: '/secure/BrowseProjects.jspa?selectedProjectType=software',
-  },
-  [ProductKey.JIRA_SERVICE_DESK]: {
-    label: 'Jira Service Desk',
-    Icon: createIcon(JiraServiceDeskIcon, { size: 'small' }),
-    href: '/secure/BrowseProjects.jspa?selectedProjectType=service_desk',
-  },
-  [ProductKey.JIRA_OPS]: {
-    label: 'Jira Ops',
-    Icon: createIcon(JiraOpsLogo, { size: 'small' }),
-    href: '/secure/BrowseProjects.jspa?selectedProjectType=ops',
-  },
-  [SINGLE_JIRA_PRODUCT]: {
-    label: 'Jira',
-    Icon: createIcon(JiraIcon, { size: 'small' }),
-    href: '/secure/MyJiraHome.jspa',
-  },
-  [ProductKey.OPSGENIE]: {
-    label: 'Opsgenie',
-    Icon: createIcon(OpsGenieIcon, { size: 'small' }),
-    href: 'https://app.opsgenie.com',
-  },
-};
-
 export const getObjectTypeLabel = (type: string): React.ReactNode => {
   return OBJECT_TYPE_TO_LABEL_MAP[type] ? (
     <FormattedMessage {...OBJECT_TYPE_TO_LABEL_MAP[type]} />
@@ -120,31 +66,19 @@ export const getObjectTypeLabel = (type: string): React.ReactNode => {
 };
 
 export const getFixedProductLinks = (params: {
-  canShowPeopleLink: boolean;
   isDiscoverMoreForEveryoneEnabled: boolean;
 }): SwitcherItemType[] => {
-  const fixedLinks = [];
-
-  if (params.canShowPeopleLink) {
-    fixedLinks.push({
-      key: 'people',
-      label: <FormattedMessage {...messages.people} />,
-      Icon: createIcon(PeopleLogo, { size: 'small' }),
-      href: `/people`,
-    });
-  }
-
-  if (params.isDiscoverMoreForEveryoneEnabled) {
-    // The discover more link href is intentionally empty to prioritise the onDiscoverMoreClicked callback
-    fixedLinks.push({
-      key: 'discover-more',
-      label: <FormattedMessage {...messages.discoverMore} />,
-      Icon: createIcon(AddIcon, { size: 'medium' }),
-      href: '',
-    });
-  }
-
-  return fixedLinks;
+  return params.isDiscoverMoreForEveryoneEnabled
+    ? [
+        {
+          // The discover more link href is intentionally empty to prioritise the onDiscoverMoreClicked callback
+          key: 'discover-more',
+          label: <FormattedMessage {...messages.discoverMore} />,
+          Icon: createIcon(AddIcon, { size: 'medium' }),
+          href: '',
+        },
+      ]
+    : [];
 };
 
 type AvailableProductDetails = Pick<
@@ -212,6 +146,14 @@ const BROWSE_APPS_URL: { [Key in Product]?: string | undefined } = {
   [Product.JIRA]: '/plugins/servlet/ac/com.atlassian.jira.emcee/discover',
   [Product.CONFLUENCE]:
     '/wiki/plugins/servlet/ac/com.atlassian.confluence.emcee/discover',
+};
+
+const TO_WORKLENS_PRODUCT_KEY: { [Key in ProductKey]: WorklensProductType } = {
+  [ProductKey.CONFLUENCE]: WorklensProductType.CONFLUENCE,
+  [ProductKey.JIRA_CORE]: WorklensProductType.JIRA_BUSINESS,
+  [ProductKey.JIRA_SERVICE_DESK]: WorklensProductType.JIRA_SERVICE_DESK,
+  [ProductKey.JIRA_SOFTWARE]: WorklensProductType.JIRA_SOFTWARE,
+  [ProductKey.OPSGENIE]: WorklensProductType.OPSGENIE,
 };
 
 interface ConnectedSite {
@@ -300,61 +242,6 @@ export const getAvailableProductLinks = (
   }).filter(link => !!link);
 };
 
-export const getProductLink = (
-  productKey: ProductKey | typeof SINGLE_JIRA_PRODUCT,
-  productLicenseInformation?: ProductLicenseInformation,
-): SwitcherItemType => {
-  const productLinkProperties = PRODUCT_DATA_MAP[productKey];
-
-  if (productKey === ProductKey.OPSGENIE && productLicenseInformation) {
-    // Prefer applicationUrl provided by license information (TCS)
-    // Fallback to hard-coded URL
-    const href = productLicenseInformation.applicationUrl
-      ? productLicenseInformation.applicationUrl
-      : productLinkProperties.href;
-
-    return { key: productKey, ...productLinkProperties, href };
-  }
-
-  return {
-    key: productKey,
-    ...productLinkProperties,
-  };
-};
-
-export const getProductIsActive = (
-  { products }: LicenseInformationResponse,
-  productKey: string,
-): boolean =>
-  products.hasOwnProperty(productKey) &&
-  products[productKey].state === ProductActivationStatus.ACTIVE;
-
-// This function will determine which product links to render based
-// on license information and if we're separating the jira products or not
-export const getLicensedProductLinks = (
-  licenseInformationData: LicenseInformationResponse,
-): SwitcherItemType[] => {
-  const majorJiraProducts = [
-    ProductKey.JIRA_SOFTWARE,
-    ProductKey.JIRA_SERVICE_DESK,
-    ProductKey.JIRA_OPS,
-  ].filter(productKey =>
-    getProductIsActive(licenseInformationData, productKey),
-  );
-  const minorJiraProducts = [ProductKey.JIRA_CORE].filter(productKey =>
-    getProductIsActive(licenseInformationData, productKey),
-  );
-
-  const jiraProducts = [...majorJiraProducts, ...minorJiraProducts];
-  const otherProducts = [ProductKey.CONFLUENCE, ProductKey.OPSGENIE].filter(
-    productKey => getProductIsActive(licenseInformationData, productKey),
-  );
-
-  return [...jiraProducts, ...otherProducts].map(productKey =>
-    getProductLink(productKey, licenseInformationData.products[productKey]),
-  );
-};
-
 export const getAdministrationLinks = (
   isAdmin: boolean,
   isDiscoverMoreForEveryoneEnabled: boolean,
@@ -376,7 +263,7 @@ export const getAdministrationLinks = (
       key: 'browse-apps',
       label: <FormattedMessage {...messages.browseApps} />,
       Icon: createIcon(MarketplaceGlyph, { size: 'medium' }),
-      href: `${emceeLink}?source=app_switcher`,
+      href: `${emceeLink}#!/discover?source=app_switcher`,
     });
   }
   if (!isDiscoverMoreForEveryoneEnabled) {
@@ -393,24 +280,43 @@ export const getAdministrationLinks = (
 const PRODUCT_RECOMMENDATION_LIMIT = 2;
 
 export const getSuggestedProductLink = (
-  licenseInformationData: LicenseInformationResponse,
+  provisionedProducts: ProvisionedProducts,
   productRecommendations: RecommendationsEngineResponse,
 ): SwitcherItemType[] => {
-  const filteredProducts = productRecommendations.filter(
-    product => !getProductIsActive(licenseInformationData, product.productKey),
+  return productRecommendations
+    .filter(legacyProduct => {
+      const productKey = TO_WORKLENS_PRODUCT_KEY[legacyProduct.productKey];
+      return !provisionedProducts[productKey];
+    })
+    .map(legacyProduct => {
+      const productKey = TO_WORKLENS_PRODUCT_KEY[legacyProduct.productKey];
+      return {
+        key: legacyProduct.productKey,
+        ...AVAILABLE_PRODUCT_DATA_MAP[productKey],
+      };
+    })
+    .slice(0, PRODUCT_RECOMMENDATION_LIMIT);
+};
+
+export const getProvisionedProducts = (
+  availableProducts: AvailableProductsResponse,
+): ProvisionedProducts => {
+  const provisionedProducts = {} as ProvisionedProducts;
+  availableProducts.sites.forEach(site =>
+    site.availableProducts.forEach(
+      product => (provisionedProducts[product.productType] = true),
+    ),
   );
-  return filteredProducts
-    .slice(0, PRODUCT_RECOMMENDATION_LIMIT)
-    .map(product => getProductLink(product.productKey));
+  return provisionedProducts;
 };
 
 export const getCustomLinkItems = (
   list: Array<CustomLink>,
-  licenseInformationData: LicenseInformationResponse,
+  currentSite: CurrentSite,
 ): SwitcherItemType[] => {
   const defaultProductCustomLinks = [
-    `${licenseInformationData.hostname}/secure/MyJiraHome.jspa`,
-    `${licenseInformationData.hostname}/wiki/`,
+    `${currentSite.url}/secure/MyJiraHome.jspa`,
+    `${currentSite.url}/wiki/`,
   ];
   return list
     .filter(
@@ -426,16 +332,20 @@ export const getCustomLinkItems = (
 
 export const getRecentLinkItems = (
   list: Array<RecentContainer>,
-  licenseInformationData: LicenseInformationResponse,
+  currentSite: CurrentSite,
 ): RecentItemType[] => {
-  const isAnyJiraProductActive =
-    getProductIsActive(licenseInformationData, ProductKey.JIRA_SOFTWARE) ||
-    getProductIsActive(licenseInformationData, ProductKey.JIRA_SERVICE_DESK) ||
-    getProductIsActive(licenseInformationData, ProductKey.JIRA_CORE) ||
-    getProductIsActive(licenseInformationData, ProductKey.JIRA_OPS);
-  const isConfluenceActive = getProductIsActive(
-    licenseInformationData,
-    ProductKey.CONFLUENCE,
+  const isAnyJiraProductActive = Boolean(
+    currentSite.products.find(
+      product =>
+        product.productType === WorklensProductType.JIRA_BUSINESS ||
+        product.productType === WorklensProductType.JIRA_SERVICE_DESK ||
+        product.productType === WorklensProductType.JIRA_SOFTWARE,
+    ),
+  );
+  const isConfluenceActive = Boolean(
+    currentSite.products.find(
+      product => product.productType === WorklensProductType.CONFLUENCE,
+    ),
   );
   return list
     .filter((recent: RecentContainer) => {