+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @atlaskit/navigation-next
 
+## 6.8.0
+
+### Minor Changes
+
+- [minor][6cc95a6c66](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6cc95a6c66):
+
+  - Support dynamic removal of contextual navigation via the `showContextualNavigation` prop
+  - Support content dataset that targets the page content element
+  - Remove re-renders occurring from the resize control hover effect
+  - Reduce re-renders occurring from the resize transition
+
+## 6.7.7
+
+### Patch Changes
+
+- [patch][c479e7eb8f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c479e7eb8f):
+
+  Fixing nav resize button render issue.
+
 ## 6.7.6
 
 ### Patch Changes